+++ app/Components/Home.js
@@ -1,4 +1,5 @@
 import React, { useState } from 'react';
+import storage from './Storage';
 import {
     StatusBar,
     StyleSheet,
@@ -6,22 +7,33 @@ import {
     View,
     TouchableOpacity,
     SafeAreaView,
-    ImageBackground,
 } from 'react-native';
 
 const scalesColors = require('../colors.json');
 
 const Home = ({route, navigation}) => {
-    const id = route.params.id;
+    var id = route.params.id;
+    
+    // offline mode tries to retrieve login info from local
+    // storage, returns to login form on fail
+    if (route.params.offlineMode) {
+        storage.load({
+            key: 'loginState',
+        }).then((local) => {
+            id = local.id;
+            console.log('Loaded local user data.');
+        }).catch((err) => {
+            Alert.alert("ERROR", err.message);
+            navigation.navigate('LoginForm');
+        });
+    }
     
     return (
         <SafeAreaView style={styles.safeArea}>
         <View style={styles.container}>
             <TouchableOpacity style={styles.buttonView}
             onPress= {() => {
-                navigation.navigate('DataEntry', {
-                    "id": id
-                });
+                navigation.navigate('DataEntry', route.params);
             }}>
                 <Text styles={styles.buttonText}>Data Entry</Text>
             </TouchableOpacity>
@@ -35,8 +47,9 @@ const Home = ({route, navigation}) => {
 
             <TouchableOpacity style= {styles.buttonView}
             onPress= {() => {
+                navigation.navigate('PrevEntries', route.params);
             }}>
-                <Text styles={styles.buttonText}>Previous Entries</Text>
+                <Text styles={styles.buttonText}>Saved Entries</Text>
             </TouchableOpacity>
 
             <TouchableOpacity style= {styles.buttonView}